Add validations and error handling in services using try catch and good development practices.
