--- src/Result/dtos/createResult.dto.ts
@@ -1,19 +1,0 @@
-import { IsInt, IsOptional } from 'class-validator';
-
-export class CreateResultDto {
-  @IsInt()
-  tournamentId: number;
-
-  @IsInt()
-  winnerId: number;
-
-  @IsInt()
-  loserId: number;
-
-  @IsInt()
-  winnerScore: number;
-
-  @IsInt()
-  loserScore: number;
-}
-
