--- src/Result/dtos/exportDto.ts
@@ -1,2 +1,0 @@
-export * from './createResult.dto';
-export * from './updateResul.dto'
