--- src/Result/dtos/updateResul.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from "@nestjs/mapped-types";
-import { CreateResultDto } from "./createResult.dto";
-
-
-
-export class UpdateMatchResultDto extends PartialType(CreateResultDto) {}
