--- src/Tournament/dtos/createtournament.dto.ts
@@ -1,16 +1,0 @@
-import { IsString, IsDate, IsOptional } from 'class-validator';
-
-export class CreateTournamentDto {
-  @IsString()
-  name: string;
-
-  @IsDate()
-  startDate: Date;
-
-  @IsDate()
-  endDate: Date;
-
-  @IsString()
-  status: string;
-}
-
