--- src/Tournament/dtos/exportDto.ts
@@ -1,2 +1,0 @@
-export * from './createtournament.dto';
-export * from './updateTournament.dto';
