--- src/Tournament/dtos/updateTournament.dto.ts
@@ -1,6 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateTournamentDto } from './createtournament.dto';
-
-
-
-export class UpdateTournamentDto extends PartialType(CreateTournamentDto) {}
