+++ src/app.module.ts
@@ -1,14 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import config from './Global/config/config';
-import { PlayerModule } from './player/module/player.module';
-import { ResultModule } from './Result/module/result.module';
-import { TournamentModule } from './Tournament/module/tournament.module';
+import { PlayerModule } from './modules/player/module/player.module';
+import { ResultModule } from './modules/Result/module/result.module';
+import { TournamentModule } from './modules/Tournament/module/tournament.module';
 import { ConfigModule } from '@nestjs/config';
-import { PrizesModule } from './prize/module/prize.module';
+import { PrizesModule } from './modules/prize/module/prize.module';
 import { ScheduleModule } from '@nestjs/schedule';
-
-
 
 @Module({
   imports: [
@@ -24,11 +22,10 @@ import { ScheduleModule } from '@nestjs/schedule';
       password: config().database.password,
       database: config().database.db,
       autoLoadEntities: true,
-      synchronize: true, 
+      synchronize: true,
       extra: {
-        ssl: true, 
+        ssl: true,
       },
-     
     }),
     ScheduleModule.forRoot(),
     PlayerModule,
