+++ src/modules/Result/dtos/exportDto.ts
@@ -1,0 +1,2 @@
+export * from './createResult.dto';
+export * from './updateResul.dto';
