+++ src/modules/Result/dtos/updateResul.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateResultDto } from './createResult.dto';
+
+export class UpdateMatchResultDto extends PartialType(CreateResultDto) {}
