+++ src/modules/Tournament/dtos/createtournament.dto.ts
@@ -1,0 +1,20 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsString, IsDate } from 'class-validator';
+
+export class CreateTournamentDto {
+  @ApiProperty()
+  @IsString()
+  name: string;
+
+  @ApiProperty()
+  @IsDate()
+  startDate: Date;
+
+  @ApiProperty()
+  @IsDate()
+  endDate: Date;
+
+  @ApiProperty()
+  @IsString()
+  status: string;
+}
