+++ src/modules/Tournament/dtos/exportDto.ts
@@ -1,0 +1,2 @@
+export * from './createtournament.dto';
+export * from './updateTournament.dto';
