+++ src/modules/Tournament/dtos/updateTournament.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateTournamentDto } from './createtournament.dto';
+
+export class UpdateTournamentDto extends PartialType(CreateTournamentDto) {}
