+++ src/modules/Tournament/entities/tournament.entity.ts
@@ -1,0 +1,40 @@
+import {
+  Entity,
+  PrimaryGeneratedColumn,
+  Column,
+  CreateDateColumn,
+  UpdateDateColumn,
+  OneToMany,
+} from 'typeorm';
+import { MatchResult } from '../../Result/entities/resul.entity';
+
+@Entity()
+export class Tournament {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  name: string;
+
+  @Column()
+  startDate: Date;
+
+  @Column()
+  endDate: Date;
+
+  @Column()
+  status: string;
+
+  @Column({ default: false })
+  isDeleted: boolean;
+
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @UpdateDateColumn()
+  updatedAt: Date;
+
+  @OneToMany(() => MatchResult, (matchResult) => matchResult.tournament)
+  matchResults: MatchResult[];
+  players: import('c:/Users/CAMILO/OneDrive/Escritorio/Backend-Nestjs/api-torneos/src/modules/player/entities/payer.entity').Player[];
+}
