+++ src/modules/player/dtos/exportDto.ts
@@ -1,0 +1,2 @@
+export * from './createPlayer.dto';
+export * from './updatePlayer.dto';
