+++ src/modules/player/dtos/updatePlayer.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreatePlayerDto } from './createPlayer.dto';
+
+export class UpdatePlayerDto extends PartialType(CreatePlayerDto) {}
