--- src/player/dtos/createPlayer.dto.ts
@@ -1,14 +1,0 @@
-import { IsString, IsEmail, IsOptional } from 'class-validator';
-
-export class CreatePlayerDto {
-  @IsString()
-  name: string;
-
-  @IsEmail()
-  email: string;
-
-  @IsOptional()
-  @IsString()
-  stats?: string;
-}
-
