--- src/player/dtos/exportDto.ts
@@ -1,2 +1,0 @@
-export * from './createPlayer.dto';
-export * from './updatePlayer.dto';
