--- src/player/dtos/updatePlayer.dto.ts
@@ -1,5 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePlayerDto } from './createPlayer.dto';
-
-
-export class UpdatePlayerDto extends PartialType(CreatePlayerDto) {}
